pyproject.toml
[project]
name = "persistent-memory-experiment"
version = "0.1.0"
description = "Persistent Memory for Frozen Encoder-Decoder LLMs: A Latent-Space Approach"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [{name = "John Doe"}]
readme = "README.md"

dependencies = [
    "torch>=2.2.0",
    "transformers>=4.40.0",
    "datasets>=2.19.0",
    "evaluate>=0.4.0",
    "sentencepiece>=0.1.99",
    "protobuf>=4.25.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "ruff>=0.3.0"]

[tool.uv]
package = false
dev-dependencies = ["pytest>=8.0.0", "ruff>=0.3.0"]

[[tool.uv.index]]
name = "pytorch-cu124"
url  = "https://download.pytorch.org/whl/cu124"
explicit = true

[tool.uv.sources]
torch = [{ index = "pytorch-cu124" }]
